+++ components/php_ext_s3_onboard_rgb/php_s3_onboard_rgb.h
@@ -1,0 +1,9 @@
+#ifndef PHP_S3_ONBOARD_RGB_H
+#define PHP_S3_ONBOARD_RGB_H
+
+#include "php.h"
+
+extern zend_module_entry s3_onboard_rgb_module_entry;
+PHP_MINIT_FUNCTION(s3_onboard_rgb);
+
+#endif /* PHP_S3_ONBOARD_RGB_H */